+++ tests/testthat/test_mutual_total.R
@@ -8,7 +8,7 @@ test_data <- data.frame(
     n = c(40, 20, 5, 1, 20, 40, 60, 80)
 )
 
-test_that("mutual works both ways around", {
+test_that("mutual M works both ways around", {
     expect_equal(
         mutual_total(test_data, "u", "g", weight = "n")[["M", "est"]],
         mutual_total(test_data, "g", "u", weight = "n")[["M", "est"]]
@@ -31,6 +31,11 @@ test_that("between + within = total", {
         mutual_total(test_data, "u", "supergroup", weight = "n")[["M", "est"]] +
             mutual_total(test_data, "u", "g", within = "supergroup", weight = "n")[["M", "est"]]
     )
+    expect_equal(
+        mutual_total(test_data, "u", "g", weight = "n")[["H", "est"]],
+        mutual_total(test_data, "u", "supergroup", weight = "n")[["H", "est"]] +
+            mutual_total(test_data, "u", "g", within = "supergroup", weight = "n")[["H", "est"]]
+    )
 })
 
 p_12 <- sum(test_data[test_data$supergroup == 12, "n"]) / sum(test_data$n)
@@ -41,11 +46,12 @@ test_that("within estimations is correct", {
             p_34 * mutual_total(test_data[test_data$supergroup == 34, ], "u", "g", weight = "n")[["M", "est"]],
         mutual_total(test_data, "u", "g", within = "supergroup", weight = "n")[["M", "est"]]
     )
+    # this decomposition does not exist for H
 })
 
 test_that("bootstrapping works", {
-    ret = mutual_total(test_data, "u", "g", weight = "n", se = TRUE, n_bootstrap = 20)
-    expect_equal(dim(ret), c(3, 3))
+    ret = mutual_total(test_data, "u", "g", weight = "n", se = TRUE, n_bootstrap = 5)
+    expect_equal(dim(ret), c(2, 3))
 })
 
 test_data <- data.frame(
@@ -55,11 +61,7 @@ test_data <- data.frame(
 )
 
 test_that("zero weights no problem", {
-    expect_equal(dim(mutual_total(test_data, "u", "g", weight = "n", se = TRUE)), c(3, 3))
-    expect_equal(dim(mutual_total(test_data, "u", "g", weight = "n")), c(3, 2))
+    expect_equal(dim(mutual_total(test_data, "u", "g", weight = "n", se = TRUE)), c(2, 3))
+    expect_equal(dim(mutual_total(test_data, "u", "g", weight = "n")), c(2, 2))
 })
 
-test_that("max bound correct", {
-    expect_equal(mutual_total(test_data, "u", "g", weight = "n")[["M", "est"]], log(2))
-    expect_equal(mutual_total(test_data, "u", "g", weight = "n")[["M_max", "est"]], log(2))
-})