+++ NEWS.md
@@ -1,6 +1,7 @@
 Development version
 
 * switch group and unit definitions, to be consistent with the literature
+* add Theil's Information Index (H) 
 
 v0.0.1
 ======